+++ app/src/main/java/com/example/mobile_programming_term_project/makeCalculationStack.java
@@ -4,6 +4,6 @@ public interface makeCalculationStack {
     boolean isEmpty();
     boolean isFull();
     void push(int item);
-    float pop();
-    float peek();
+    int pop();
+    int peek();
 }
